Split plots module into plotting subpackage
