+++ dataviews/__init__.py
@@ -17,8 +17,8 @@ from .ndmapping import * # pyflakes:ignore (API import)
 
 def public(obj):
     if not isinstance(obj, type): return False
-    baseclasses = [NdMapping, View, Dimension]
+    baseclasses = [NdMapping, View, Dimension, Overlay]
     return any([issubclass(obj, bc) for bc in baseclasses])
 
 _public = list(set([_k for _k, _v in locals().items() if public(_v)]))
-__all__ = _public + ["boundingregion", "ipython", "plots", "sheetcoords" ]
+__all__ = _public + ["boundingregion", "ipython", "plotting", "sheetcoords" ]
